+++ mayan/apps/sources/tasks.py
@@ -45,13 +45,8 @@ def task_check_interval_source(source_id, test=False):
             logger.error(
                 'Error processing source id: %s; %s', source_id, exception
             )
-            source.logs.create(
-                message=_('Error processing source: %s') % exception
-            )
             if settings.DEBUG:
                 raise
-        else:
-            source.logs.all().delete()
         finally:
             lock.release()